update to 0.4.1

+++ lib/backgrounder/workers/store_asset.rb
@@ -8,11 +8,11 @@ module CarrierWave
       def perform(*args)
         record = super(*args)
 
-        if record.send(:"#{column}_tmp")
+        if record && record.send(:"#{column}_tmp")
           store_directories(record)
           record.send :"process_#{column}_upload=", true
           record.send :"#{column}_tmp=", nil
-          record.send :"#{column}_processing=", nil if record.respond_to?(:"#{column}_processing")
+          record.send :"#{column}_processing=", false if record.respond_to?(:"#{column}_processing")
           File.open(cache_path) { |f| record.send :"#{column}=", f }
           if record.save!
             FileUtils.rm_r(tmp_directory, :force => true)
